src/backend/requirements.txt
fastapi
uvicorn
httpx
jinja2
asyncpg
python-dotenv
PyJWT
requests
sqlalchemy